+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 February 2023
 
+- Add 'm' keyboard shortcut for toggling move unit interaction.
 - Add context menu to map.
 - Make chart mode sidebar resizable.
 - Allow adding an arbitrary number of sides and root units in the 'create new scenario form'.